requirements.txt
requests==2.32.3
pyyaml==6.0.2
python-telegram-bot==21.5